Renamed script.js -> main.js and added scripts.js file
