+++ src/js/components/GeneralComponents.js
@@ -32,14 +32,4 @@ const TextBox = (textBoxClass, size, holderText, keyUpFunction) => {
     return e('input', {className: textBoxClass, size: size, placeholder: holderText, onKeyUp: keyUpFunction});
 }
 
-const ButtonEnable = (textBoxClass, buttonClass) => {
-    let form = document.getElementsByClassName(textBoxClass) && document.getElementsByClassName(textBoxClass)[0];
-    let button = document.getElementsByClassName(buttonClass) && document.getElementsByClassName(buttonClass)[0];
-    if(form.value) {
-        button.disabled = false;
-    } else {
-        button.disabled = true;
-    }
-}
-
-export {Header, Mismatch, Pending, Error, Button, TextBox, ButtonEnable};
+export {Header, Mismatch, Pending, Error, Button, TextBox};
